move configuration values to a config/system_pay.yml file; add railtie to load yml on startup; load form_helper automatically; update Readme accordingly
