--- init.rb
@@ -1,2 +1,0 @@
-require "system_pay"
-require "syetem_pay/form_helper"
