+++ lib/system_pay/form_helper.rb
@@ -1,13 +1,11 @@
-module SystemPay::FormHelper
-
-  def system_pay_hidden_fields(system_pay)
-  
-    res = "\n"  
-    system_pay.params.each{|key, value|
-      res << hidden_field_tag(key, value) << "\n"
-    }
-    res.html_safe
-  
+module SystemPay
+  module FormHelper
+    def system_pay_hidden_fields(system_pay)
+      res = "\n"  
+      system_pay.params.each do |key, value|
+        res << hidden_field_tag(key, value) << "\n"
+      end
+      res.html_safe
+    end
   end
-
 end
