+++ lib/system_pay/version.rb
@@ -1,3 +1,3 @@
-class SystemPay
+module SystemPay
   VERSION = '0.0.5'
 end
