+++ system_pay.gemspec
@@ -14,6 +14,8 @@ Gem::Specification.new do |s|
   s.email     = ['sylvain@imenlo.com']
   s.homepage  = 'https://github.com/iMenlo/system_pay'
 
+  s.add_dependency "rails", "~> 3.0"
+
   s.add_development_dependency "active_support", "~> 3.0.0"
   s.add_development_dependency 'rake', '~> 0.9.2'
   s.add_development_dependency 'rspec', '~> 2.6.0'
